requirements.txt
aioredis==1.2.0
amqp==2.3.2
asgiref==2.3.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.11.1
Automat==0.7.0
bcrypt==3.1.4
billiard==3.5.0.4
certifi==2018.10.15
cffi==1.11.5
channels==2.1.5
channels-redis==2.3.1
chardet==3.0.4
ConcurrentLogHandler==0.9.1
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
daphne==2.2.3
defusedxml==0.5.0
Django==2.1.3
django-allauth==0.38.0
django-configurations==2.1
django-cors-headers==2.4.0
django-environ==0.4.5
django-rest-auth==0.9.3
django-rest-knox==3.6.0
django-sslserver==0.20
django-timezone-field==3.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
dnspython==1.15.0
ephem==3.7.6.0
eventlet==0.24.1
Fabric3==1.14.post1
freeze==1.0.10
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
hiredis==0.2.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.1.0
monotonic==1.5
msgpack==0.5.6
oauthlib==2.1.0
olefile==0.46
paramiko==2.4.2
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.6.4
PyNaCl==1.3.0
pyOpenSSL==18.0.0
python-crontab==2.3.5
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
requests==2.20.0
requests-oauthlib==1.0.0
service-identity==18.1.0
six==1.11.0
Twisted==18.9.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.24.1
vine==1.1.4
whitenoise==4.1.2
zope.interface==4.6.0
